Split reactive and non reactive methods in distinct classes
