+++ jhipster-framework/src/test/java/io/github/jhipster/web/util/ResponseUtilTest.java
@@ -25,7 +25,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.server.ResponseStatusException;
-import reactor.core.publisher.Mono;
 
 import java.util.Optional;
 
@@ -38,16 +37,12 @@ public class ResponseUtilTest {
 
     private Optional<Integer> optionalYes;
     private Optional<Integer> optionalNo;
-    private Mono<Integer> monoYes;
-    private Mono<Integer> monoNo;
     private HttpHeaders headers;
 
     @BeforeEach
     public void setup() {
         optionalYes = Optional.of(42);
         optionalNo = Optional.empty();
-        monoYes = Mono.just(42);
-        monoNo = Mono.empty();
         headers = new HttpHeaders();
         headers.add(HEADER_NAME, HEADER_VALUE);
     }
@@ -80,33 +75,4 @@ public class ResponseUtilTest {
         assertThatExceptionOfType(ResponseStatusException.class).isThrownBy(() -> ResponseUtil.wrapOrNotFound(optionalNo, headers));
     }
 
-    @Test
-    public void testMonoYesWithoutHeaders() {
-        ResponseEntity<Integer> response = ResponseUtil.wrapOrNotFound(monoYes).block();
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(response.getBody()).isEqualTo(42);
-        assertThat(response.getHeaders()).isEmpty();
-    }
-
-    @Test
-    public void testMonoNoWithoutHeaders() {
-        Mono<ResponseEntity<Integer>> response = ResponseUtil.wrapOrNotFound(monoNo);
-        assertThatExceptionOfType(ResponseStatusException.class).isThrownBy(response::block);
-    }
-
-    @Test
-    public void testMonoYesWithHeaders() {
-        ResponseEntity<Integer> response = ResponseUtil.wrapOrNotFound(monoYes, headers).block();
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(response.getBody()).isEqualTo(42);
-        assertThat(response.getHeaders()).hasSize(1);
-        assertThat(response.getHeaders().get(HEADER_NAME)).hasSize(1);
-        assertThat(response.getHeaders().get(HEADER_NAME).get(0)).isEqualTo(HEADER_VALUE);
-    }
-
-    @Test
-    public void testMonoNoWithHeaders() {
-        Mono<ResponseEntity<Integer>> response = ResponseUtil.wrapOrNotFound(monoNo, headers);
-        assertThatExceptionOfType(ResponseStatusException.class).isThrownBy(response::block);
-    }
 }
